move is_inside_subpath function to Subpath struct method
