+++ libraries/bezier-rs/src/subpath/solvers.rs
@@ -1,6 +1,6 @@
 use super::*;
 use crate::consts::MAX_ABSOLUTE_DIFFERENCE;
-use crate::utils::{compute_circular_subpath_details, line_intersection, SubpathTValue};
+use crate::utils::{compute_circular_subpath_details, is_rectangle_inside_other, line_intersection, SubpathTValue};
 use crate::TValue;
 
 use glam::{DAffine2, DMat2, DVec2};
@@ -235,6 +235,37 @@ impl<PointId: crate::Identifier> Subpath<PointId> {
 		}
 
 		false
+	}
+
+	/// Returns `true` if this subpath is completely inside the other subpath.
+	pub fn is_inside_subpath(&self, other: &Subpath<PointId>, error: Option<f64>, minimum_separation: Option<f64>) -> bool {
+		// Eliminate this subpath if its bounding box is not completely inside other subpath bounding box
+		if !self.is_empty() && !other.is_empty() {
+			let inner_bbox = self.bounding_box().unwrap();
+			let outer_bbox = other.bounding_box().unwrap();
+			// Reasoning:
+			// (min x, min y) of inner subpath is less or equal to the outer (min x, min y) or
+			// (min x, min y) of inner subpath is more or equal to outer (max x, mix y) then the inner is intersecting or is outside the outer. (same will be true for (max x, max y))
+			if !is_rectangle_inside_other(inner_bbox, outer_bbox) {
+				return false;
+			}
+		};
+
+		// Eliminate this if any of its the subpath's anchors is outside other's subpath
+		for anchors in self.anchors() {
+			if !other.contains_point(anchors) {
+				return false;
+			}
+		}
+
+		// Eliminate this if its subpath is intersecting wih the other's subpath
+		if !self.subpath_intersections(&other, error, minimum_separation).is_empty() {
+			return false;
+		}
+
+		// here (1) this subpath bbox is inside other bbox, (2) its anchors are inside other subpath and (3) it is not intersecting with other subpath.
+		// hence this this subpath is completely inside given other subpath.
+		true
 	}
 
 	/// Returns a normalized unit vector representing the tangent on the subpath based on the parametric `t`-value provided.
@@ -875,6 +906,28 @@ mod tests {
 	}
 
 	// TODO: add more intersection tests
+
+	#[test]
+	fn is_inside_subpath() {
+		let lasso_polygon = [DVec2::new(100., 100.), DVec2::new(500., 100.), DVec2::new(500., 500.), DVec2::new(100., 500.)].to_vec();
+		let lasso_polygon = Subpath::from_anchors_linear(lasso_polygon, true);
+
+		let curve = Bezier::from_quadratic_dvec2(DVec2::new(189., 289.), DVec2::new(9., 286.), DVec2::new(45., 410.));
+		let curve_intersecting = Subpath::<EmptyId>::from_bezier(&curve);
+		assert_eq!(curve_intersecting.is_inside_subpath(&lasso_polygon, None, None), false);
+
+		let curve = Bezier::from_quadratic_dvec2(DVec2::new(115., 37.), DVec2::new(51.4, 91.8), DVec2::new(76.5, 242.));
+		let curve_outside = Subpath::<EmptyId>::from_bezier(&curve);
+		assert_eq!(curve_outside.is_inside_subpath(&lasso_polygon, None, None), false);
+
+		let curve = Bezier::from_cubic_dvec2(DVec2::new(210.1, 133.5), DVec2::new(150.2, 436.9), DVec2::new(436., 285.), DVec2::new(247.6, 240.7));
+		let curve_inside = Subpath::<EmptyId>::from_bezier(&curve);
+		assert_eq!(curve_inside.is_inside_subpath(&lasso_polygon, None, None), true);
+
+		let line = Bezier::from_linear_dvec2(DVec2::new(101., 101.5), DVec2::new(150.2, 499.));
+		let line_inside = Subpath::<EmptyId>::from_bezier(&line);
+		assert_eq!(line_inside.is_inside_subpath(&lasso_polygon, None, None), true);
+	}
 
 	#[test]
 	fn round_join_counter_clockwise_rotation() {
